add delete author

+++ src/graphQL/mutations/deleteBook.mutation.jsx
@@ -1,7 +1,7 @@
 import { gql } from "@apollo/client";
 
 export const DELETE_BOOK = gql`
-  mutation deleteBook($id: Int!) {
+  mutation deleteBook($id: Float!) {
     deleteBookById(input: { bookId: $id })
   }
 `;
